separate and document FlyAwayOnIdle
